+++ database/upgrade/upgrade4.0.0.sql
@@ -7,6 +7,11 @@
 START TRANSACTION;
 
 RENAME TABLE myems_system_db.tbl_energy_storage_power_stations TO myems_system_db.tbl_energy_storage_containers;
+ALTER TABLE myems_system_db.tbl_energy_storage_containers DROP COLUMN address;
+ALTER TABLE myems_system_db.tbl_energy_storage_containers DROP COLUMN postal_code;
+ALTER TABLE myems_system_db.tbl_energy_storage_containers DROP COLUMN latitude;
+ALTER TABLE myems_system_db.tbl_energy_storage_containers DROP COLUMN longitude;
+
 DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_batteries` ;
 DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_commands` ;
 DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_grids` ;
@@ -109,12 +114,6 @@ CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_containers_sens
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_energy_storage_containers_sensors_index_1` ON  `myems_system_db`.`tbl_energy_storage_containers_sensors` (`energy_storage_container_id`);
 
-
--- ---------------------------------------------------------------------------------------------------------------------
--- Table `myems_system_db`.`tbl_energy_storage_power_stations`
--- ---------------------------------------------------------------------------------------------------------------------
-DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations` ;
-
 CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(255) NOT NULL,
@@ -131,11 +130,6 @@ CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations`
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_energy_storage_power_stations_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations` (`name`);
 
--- ---------------------------------------------------------------------------------------------------------------------
--- Table `myems_system_db`.`tbl_energy_storage_power_stations_containers`
--- ---------------------------------------------------------------------------------------------------------------------
-DROP TABLE IF EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_containers` ;
-
 CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_containers` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `energy_storage_power_station_id` BIGINT NOT NULL,
@@ -151,6 +145,6 @@ CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_energy_storage_power_stations_
 CREATE INDEX `tbl_energy_storage_power_stations_users_index_1` ON  `myems_system_db`.`tbl_energy_storage_power_stations_users` (`energy_storage_power_station_id`);
 
 -- UPDATE VERSION NUMBER
-UPDATE `myems_system_db`.`tbl_versions` SET version='4.0.0RC', release_date='2024-01-01' WHERE id=1;
+UPDATE `myems_system_db`.`tbl_versions` SET version='4.0.0', release_date='2024-01-13' WHERE id=1;
 
 COMMIT;