+++ myems-api/core/meter.py
@@ -1,4 +1,5 @@
 import uuid
+from datetime import datetime, timezone, timedelta
 import falcon
 import mysql.connector
 import simplejson as json
@@ -1329,3 +1330,390 @@ class MeterCommandItem:
         cnx.close()
 
         resp.status = falcon.HTTP_204
+
+
+class MeterExport:
+    @staticmethod
+    def __init__():
+        """Initializes MeterExport"""
+        pass
+
+    @staticmethod
+    def on_options(req, resp, id_):
+        resp.status = falcon.HTTP_200
+
+    @staticmethod
+    def on_get(req, resp, id_):
+        if 'API-KEY' not in req.headers or \
+                not isinstance(req.headers['API-KEY'], str) or \
+                len(str.strip(req.headers['API-KEY'])) == 0:
+            access_control(req)
+        else:
+            api_key_control(req)
+        if not id_.isdigit() or int(id_) <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_METER_ID')
+
+        cnx = mysql.connector.connect(**config.myems_system_db)
+        cursor = cnx.cursor()
+
+        query = (" SELECT id, name, uuid "
+                 " FROM tbl_energy_categories ")
+        cursor.execute(query)
+        rows_energy_categories = cursor.fetchall()
+
+        energy_category_dict = dict()
+        if rows_energy_categories is not None and len(rows_energy_categories) > 0:
+            for row in rows_energy_categories:
+                energy_category_dict[row[0]] = {"id": row[0],
+                                                "name": row[1],
+                                                "uuid": row[2]}
+
+        query = (" SELECT id, name, uuid "
+                 " FROM tbl_cost_centers ")
+        cursor.execute(query)
+        rows_cost_centers = cursor.fetchall()
+
+        cost_center_dict = dict()
+        if rows_cost_centers is not None and len(rows_cost_centers) > 0:
+            for row in rows_cost_centers:
+                cost_center_dict[row[0]] = {"id": row[0],
+                                            "name": row[1],
+                                            "uuid": row[2]}
+
+        query = (" SELECT id, name, uuid "
+                 " FROM tbl_energy_items ")
+        cursor.execute(query)
+        rows_energy_items = cursor.fetchall()
+
+        energy_item_dict = dict()
+        if rows_energy_items is not None and len(rows_energy_items) > 0:
+            for row in rows_energy_items:
+                energy_item_dict[row[0]] = {"id": row[0],
+                                            "name": row[1],
+                                            "uuid": row[2]}
+
+        query = (" SELECT id, name, uuid "
+                 " FROM tbl_meters ")
+        cursor.execute(query)
+        rows_master_meters = cursor.fetchall()
+
+        master_meter_dict = dict()
+        if rows_master_meters is not None and len(rows_master_meters) > 0:
+            for row in rows_master_meters:
+                master_meter_dict[row[0]] = {"id": row[0],
+                                             "name": row[1],
+                                             "uuid": row[2]}
+
+        query = (" SELECT id, name, uuid, energy_category_id, "
+                 "        is_counted, hourly_low_limit, hourly_high_limit, "
+                 "        cost_center_id, energy_item_id, master_meter_id, description "
+                 " FROM tbl_meters "
+                 " WHERE id = %s ")
+        cursor.execute(query, (id_,))
+        row = cursor.fetchone()
+        cursor.close()
+        cnx.close()
+
+        if row is None:
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.METER_NOT_FOUND')
+        else:
+            energy_category = energy_category_dict.get(row[3], None)
+            cost_center = cost_center_dict.get(row[7], None)
+            energy_item = energy_item_dict.get(row[8], None)
+            master_meter = master_meter_dict.get(row[9], None)
+            meta_result = {"name": row[1],
+                           "uuid": row[2],
+                           "energy_category": energy_category,
+                           "is_counted": True if row[4] else False,
+                           "hourly_low_limit": row[5],
+                           "hourly_high_limit": row[6],
+                           "cost_center": cost_center,
+                           "energy_item": energy_item,
+                           "master_meter": master_meter,
+                           "description": row[10]}
+
+        resp.text = json.dumps(meta_result)
+
+
+class MeterImport:
+    @staticmethod
+    def __init__():
+        """Initializes MeterImport"""
+        pass
+
+    @staticmethod
+    def on_options(req, resp):
+        resp.status = falcon.HTTP_200
+
+    @staticmethod
+    @user_logger
+    def on_post(req, resp):
+        """Handles POST requests"""
+        admin_control(req)
+        try:
+            raw_json = req.stream.read().decode('utf-8')
+        except Exception as ex:
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                   title='API.BAD_REQUEST',
+                                   description='API.FAILED_TO_READ_REQUEST_STREAM')
+
+        new_values = json.loads(raw_json)
+
+        if 'name' not in new_values.keys() or \
+                not isinstance(new_values['name'], str) or \
+                len(str.strip(new_values['name'])) == 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_METER_NAME')
+        timezone_offset = int(config.utc_offset[1:3]) * 60 + int(config.utc_offset[4:6])
+        if config.utc_offset[0] == '-':
+            timezone_offset = -timezone_offset
+        name = str.strip(new_values['name']) + \
+            (datetime.utcnow() + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds')
+
+        if 'energy_category' not in new_values.keys() or \
+            'id' not in new_values['energy_category'].keys() or \
+            not isinstance(new_values['energy_category']['id'], int) or \
+                new_values['energy_category']['id'] <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_ENERGY_CATEGORY_ID')
+        energy_category_id = new_values['energy_category']['id']
+
+        if 'is_counted' not in new_values.keys() or \
+                not isinstance(new_values['is_counted'], bool):
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_IS_COUNTED_VALUE')
+        is_counted = new_values['is_counted']
+
+        if 'hourly_low_limit' not in new_values.keys() or \
+                not (isinstance(new_values['hourly_low_limit'], float) or
+                     isinstance(new_values['hourly_low_limit'], int)):
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_HOURLY_LOW_LIMIT_VALUE')
+        hourly_low_limit = new_values['hourly_low_limit']
+
+        if 'hourly_high_limit' not in new_values.keys() or \
+                not (isinstance(new_values['hourly_high_limit'], float) or
+                     isinstance(new_values['hourly_high_limit'], int)):
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_HOURLY_HIGH_LIMIT_VALUE')
+        hourly_high_limit = new_values['hourly_high_limit']
+
+        if 'cost_center' not in new_values.keys() or \
+            new_values['cost_center'] is None or \
+            'id' not in new_values['cost_center'].keys() or \
+                not isinstance(new_values['cost_center']['id'], int) or \
+                new_values['cost_center']['id'] <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_COST_CENTER_ID')
+        cost_center_id = new_values['cost_center']['id']
+
+        if 'energy_item' in new_values.keys() and \
+            new_values['energy_item'] is not None and \
+                'id' in new_values['energy_item'].keys() and \
+                new_values['energy_item']['id'] is not None:
+            if not isinstance(new_values['energy_item']['id'], int) or \
+                    new_values['energy_item']['id'] <= 0:
+                raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                       description='API.INVALID_ENERGY_ITEM_ID')
+            energy_item_id = new_values['energy_item']['id']
+        else:
+            energy_item_id = None
+
+        if 'master_meter' in new_values.keys() and \
+            new_values['master_meter'] is not None and \
+                'id' in new_values['master_meter'].keys():
+            if not isinstance(new_values['master_meter']['id'], int) or \
+                    new_values['master_meter']['id'] <= 0:
+                raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                       description='API.INVALID_MASTER_METER_ID')
+            master_meter_id = new_values['master_meter']['id']
+        else:
+            master_meter_id = None
+
+        if 'description' in new_values.keys() and \
+                new_values['description'] is not None and \
+                len(str(new_values['description'])) > 0:
+            description = str.strip(new_values['description'])
+        else:
+            description = None
+
+        cnx = mysql.connector.connect(**config.myems_system_db)
+        cursor = cnx.cursor()
+
+        cursor.execute(" SELECT name "
+                       " FROM tbl_meters "
+                       " WHERE name = %s ", (name,))
+        if cursor.fetchone() is not None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.METER_NAME_IS_ALREADY_IN_USE')
+
+        cursor.execute(" SELECT name "
+                       " FROM tbl_energy_categories "
+                       " WHERE id = %s ",
+                       (energy_category_id,))
+        if cursor.fetchone() is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.ENERGY_CATEGORY_NOT_FOUND')
+
+        cursor.execute(" SELECT name "
+                       " FROM tbl_cost_centers "
+                       " WHERE id = %s ",
+                       (new_values['cost_center']['id'],))
+        row = cursor.fetchone()
+        if row is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.COST_CENTER_NOT_FOUND')
+
+        if energy_item_id is not None:
+            cursor.execute(" SELECT name, energy_category_id "
+                           " FROM tbl_energy_items "
+                           " WHERE id = %s ",
+                           (energy_item_id,))
+            row = cursor.fetchone()
+            if row is None:
+                cursor.close()
+                cnx.close()
+                raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                       description='API.ENERGY_ITEM_NOT_FOUND')
+            else:
+                if row[1] != energy_category_id:
+                    cursor.close()
+                    cnx.close()
+                    raise falcon.HTTPError(status=falcon.HTTP_404, title='API.BAD_REQUEST',
+                                           description='API.ENERGY_ITEM_DOES_NOT_BELONG_TO_ENERGY_CATEGORY')
+
+        if master_meter_id is not None:
+            cursor.execute(" SELECT name, energy_category_id "
+                           " FROM tbl_meters "
+                           " WHERE id = %s ",
+                           (master_meter_id,))
+            row = cursor.fetchone()
+            if row is None:
+                cursor.close()
+                cnx.close()
+                raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                       description='API.MASTER_METER_NOT_FOUND')
+            else:
+                if row[1] != energy_category_id:
+                    cursor.close()
+                    cnx.close()
+                    raise falcon.HTTPError(status=falcon.HTTP_404, title='API.BAD_REQUEST',
+                                           description='API.MASTER_METER_DOES_NOT_BELONG_TO_SAME_ENERGY_CATEGORY')
+
+        add_values = (" INSERT INTO tbl_meters "
+                      "    (name, uuid, energy_category_id, is_counted, hourly_low_limit, hourly_high_limit,"
+                      "     cost_center_id, energy_item_id, master_meter_id, description) "
+                      " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
+        cursor.execute(add_values, (name,
+                                    str(uuid.uuid4()),
+                                    energy_category_id,
+                                    is_counted,
+                                    hourly_low_limit,
+                                    hourly_high_limit,
+                                    cost_center_id,
+                                    energy_item_id,
+                                    master_meter_id,
+                                    description))
+        new_id = cursor.lastrowid
+        cnx.commit()
+        cursor.close()
+        cnx.close()
+
+        resp.status = falcon.HTTP_201
+        resp.location = '/meters/' + str(new_id)
+
+
+class MeterClone:
+    @staticmethod
+    def __init__():
+        """Initializes Class"""
+        pass
+
+    @staticmethod
+    def on_options(req, resp, id_):
+        resp.status = falcon.HTTP_200
+
+    @staticmethod
+    @user_logger
+    def on_post(req, resp, id_):
+        """Handles POST requests"""
+        admin_control(req)
+        if not id_.isdigit() or int(id_) <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_METER_ID')
+
+        cnx = mysql.connector.connect(**config.myems_system_db)
+        cursor = cnx.cursor()
+
+        query = (" SELECT id, name, uuid "
+                 " FROM tbl_meters ")
+        cursor.execute(query)
+        rows_master_meters = cursor.fetchall()
+
+        master_meter_dict = dict()
+        if rows_master_meters is not None and len(rows_master_meters) > 0:
+            for row in rows_master_meters:
+                master_meter_dict[row[0]] = {"id": row[0],
+                                             "name": row[1],
+                                             "uuid": row[2]}
+
+        query = (" SELECT id, name, uuid, energy_category_id, "
+                 "        is_counted, hourly_low_limit, hourly_high_limit, "
+                 "        cost_center_id, energy_item_id, master_meter_id, description "
+                 " FROM tbl_meters "
+                 " WHERE id = %s ")
+        cursor.execute(query, (id_,))
+        row = cursor.fetchone()
+
+        if row is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.METER_NOT_FOUND')
+        else:
+            meta_result = {"id": row[0],
+                           "name": row[1],
+                           "uuid": row[2],
+                           "energy_category_id": row[3],
+                           "is_counted": row[4],
+                           "hourly_low_limit": row[5],
+                           "hourly_high_limit": row[6],
+                           "cost_center_id": row[7],
+                           "energy_item_id": row[8],
+                           "master_meter_id": row[9],
+                           "description": row[10]}
+        timezone_offset = int(config.utc_offset[1:3]) * 60 + int(config.utc_offset[4:6])
+        if config.utc_offset[0] == '-':
+            timezone_offset = -timezone_offset
+        new_name = str.strip(meta_result['name']) + \
+            (datetime.utcnow() + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds')
+
+        add_values = (" INSERT INTO tbl_meters "
+                      "    (name, uuid, energy_category_id, is_counted, hourly_low_limit, hourly_high_limit,"
+                      "     cost_center_id, energy_item_id, master_meter_id, description) "
+                      " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
+        cursor.execute(add_values, (new_name,
+                                    str(uuid.uuid4()),
+                                    meta_result['energy_category_id'],
+                                    meta_result['is_counted'],
+                                    meta_result['hourly_low_limit'],
+                                    meta_result['hourly_high_limit'],
+                                    meta_result['cost_center_id'],
+                                    meta_result['energy_item_id'],
+                                    meta_result['master_meter_id'],
+                                    meta_result['description']))
+        new_id = cursor.lastrowid
+        cnx.commit()
+        cursor.close()
+        cnx.close()
+
+        resp.status = falcon.HTTP_201
+        resp.location = '/meters/' + str(new_id)